Bug 987311 - 6/6 - Convert the rest of CompositorTypes.h to typed enums, final manual changes - r=nical

+++ gfx/layers/composite/CompositableHost.cpp
@@ -159,7 +159,7 @@ CompositableHost::AddMaskEffect(EffectChain& aEffects,
                                              source->GetSize(),
                                              aTransform);
   effect->mIs3D = aIs3D;
-  aEffects.mSecondaryEffects[EffectTypes::MASK] = effect;
+  aEffects.mSecondaryEffects[size_t(EffectTypes::MASK)] = effect;
   return true;
 }
